Add Codepipeline

+++ models/visit.go
@@ -7,7 +7,7 @@ import (
 // Visit model
 type Visit struct {
 	ObjectType string    `gorm:"not null" json:"obj_type"`
-	ObjectID   int       `gorm:"not null" json:"obj_id"`
+	ObjectID   int64     `gorm:"not null" json:"obj_id"`
 	Type       string    `json:"type"`
 	IP         string    `json:"ip"`
 	UserAgent  string    `json:"user_agent"`
